+++ climpred/classes.py
@@ -617,21 +617,33 @@ class PerfectModelEnsemble(PredictionEnsemble):
         return self._datasets['control']
 
     def verify(
-        self, comparison=None, dim=None, metric=None, reference=None, **metric_kwargs,
+        self, metric=None, comparison=None, dim=None, reference=None, **metric_kwargs,
     ):
-        """Verifies the initialized ensemble against the control run.
+        """Verify initialized predictions against a configuration of other ensemble members.
+
+        .. note::
+            The configuration of the other ensemble members is based off of the
+            ``comparison`` keyword argument.
 
         Args:
-            comparison (str): How to compare the climate prediction ensemble to the
-                control.
-            dim (str, list of str): dimension(s) to apply metric over.
-            metric (str or Metric): Metric to apply in the comparison.
-            reference (str, list of str): Reference forecasts to compute metric
-                with.
-            **metric_kwargs (optional): Arguments passed to `metric`.
+            metric (str or Metric): Metric to apply in the comparison. See
+                `metrics </metrics.html>`_.
+            comparison (str): How to compare the initialized prediction ensemble with
+                itself, see `comparisons </comparisons.html>`_.
+            dim (str, list of str): Dimension(s) over which to apply ``metric``.
+                ``dim`` is passed on to xskillscore.{metric} and includes xskillscore's
+                ``member_dim``. ``dim`` should contain ``member`` when ``comparison``
+                is probabilistic but should not contain ``member`` when
+                ``comparison=e2c``. Defaults to ``None`` meaning that all dimensions
+                other than ``lead`` are reduced.
+            reference (str, list of str): Type of reference forecasts with which to
+                verify. One or more of ['persistence', 'historical', 'uninitialized'].
+            **metric_kwargs (optional): Arguments passed to ``metric``.
 
-        Returns: Dataset of comparison results with `skill` dimension for difference
-            references compared against.
+        Returns:
+            Dataset of comparison results with ``skill`` dimension for verification
+            results for the initialized ensemble (``init``) and any reference forecasts
+            verified.
         """
         has_dataset(self._datasets['control'], 'control', 'compute a metric')
         input_dict = {
@@ -681,17 +693,30 @@ class PerfectModelEnsemble(PredictionEnsemble):
         return all_skills.squeeze()
 
     def compute_uninitialized(
-        self, comparison=None, dim=None, metric=None, **metric_kwargs
+        self, metric=None, comparison=None, dim=None, **metric_kwargs
     ):
-        """Compares the bootstrapped uninitialized run to the control run.
+        """Verify the bootstrapped uninitialized run against a configuration of other
+        ensemble members.
+
+        .. note::
+            The configuration of the other ensemble members is based off of the
+            ``comparison`` keyword argument.
 
         Args:
-            comparison (str): How to compare to the control run.
-            dim (str, list of str): Dimension(s) to apply metric over.
-            metric (str): Metric to apply in the comparison.
+            metric (str): Metric to apply in the comparison. See
+                `metrics </metrics.html>`_.
+            comparison (str): How to compare the uninitialized against itself, see
+                `comparisons </comparisons.html>`_.
+            dim (str, list of str): Dimension(s) over which to apply metric.
+                ``dim`` is passed on to xskillscore.{metric} and includes xskillscore's
+                ``member_dim``. ``dim`` should contain ``member`` when ``comparison``
+                is probabilistic but should not contain ``member`` when
+                ``comparison=e2c``. Defaults to ``None``, meaning that all dimensions
+                other than ``lead`` are reduced.
             **metric_kwargs (optional): Arguments passed to ``metric``.
 
-        Returns: Result of the comparison as a Dataset.
+        Returns:
+            Result of the comparison as a Dataset.
         """
         has_dataset(
             self._datasets['uninitialized'],
@@ -716,15 +741,14 @@ class PerfectModelEnsemble(PredictionEnsemble):
         return res
 
     def compute_persistence(self, metric=None):
-        """Compute a simple persistence forecast for the control run.
+        """Verify a simple persistence forecast of the control run against itself.
 
         Args:
-            metric (str): Metric to apply to the persistence forecast.
+            metric (str): Metric to use when comparing the persistence forecast to the
+            control run. See `metrics </metrics.html>`_.
 
         Returns:
-            Dataset of persistence forecast results (if ``refname`` is declared),
-            or dictionary of Datasets with keys corresponding to verification data
-            name.
+            Dataset of persistence forecast results.
 
         Reference:
             * Chapter 8 (Short-Term Climate Prediction) in
@@ -751,42 +775,50 @@ class PerfectModelEnsemble(PredictionEnsemble):
 
     def bootstrap(
         self,
-        iterations,
-        comparison=None,
-        dim=['init', 'member'],
         metric=None,
+        comparison=None,
+        dim=None,
+        iterations=5,
         sig=95,
         pers_sig=None,
         **metric_kwargs,
     ):
-        """Bootstrap ensemble PerfectModelEnsemble with replacement according to
-        Goddard et al. 2013.
+        """Bootstrap with replacement according to Goddard et al. 2013.
 
         Args:
-            iterations (int): Number of resampling iterations for bootstrapping
-                with replacement.
-            comparison (str): Comparison style for bootstrapping. One of ['m2c',
-                'm2m', 'e2c', 'm2e'].
-            dim (str, list of str, optional): Dimension(s) to apply metric over.
-                Defaults to ['init', 'member'].
-            metric (str): Metric to apply for bootstrapping.
-            sig (int, optional): Significance level for uninitialized and initialized
-                comparison. Defaults to 95.
+            metric (str): Metric to verify bootstrapped skill, see
+                `metrics </metrics.html>`_.
+            comparison (str): Comparison passed to verify, see
+                `comparisons </comparisons.html>`_.
+            dim (str, list of str): Dimension(s) over which to apply metric.
+                ``dim`` is passed on to xskillscore.{metric} and includes xskillscore's
+                ``member_dim``. ``dim`` should contain ``member`` when ``comparison``
+                is probabilistic but should not contain ``member`` when
+                ``comparison=e2c``. Defaults to ``None`` meaning that all dimensions
+                other than ``lead`` are reduced.
+            iterations (int): Number of resampling iterations for bootstrapping with
+                replacement. Recommended >= 500.
+            sig (int, default 95): Significance level in percent for deciding whether
+                uninitialized and persistence beat initialized skill.
             pers_sig (int): If not ``None``, the separate significance level for
                 persistence. Defaults to ``None``, or the same significance as ``sig``.
-            **metric_kwargs (optional): Arguments passed to `metric`.
+            **metric_kwargs (optional): arguments passed to ``metric``.
 
         Returns:
-            Dictionary of Datasets for each variable applied to with the
-            following variables:
-                * init_ci: confidence levels of init_skill.
-                * uninit_ci: confidence levels of uninit_skill.
-                * pers_ci: confidence levels of pers_skill.
-                * p_uninit_over_init: p value of the hypothesis that the
+            xr.Datasets: with dimensions ``result`` (holding ``skill``, ``p``,
+            ``low_ci`` and ``high_ci``) and ``kind`` (holding ``init``, ``pers`` and
+            ``uninit``):
+                * result='skill', kind='init':
+                    mean initialized skill
+                * result='skill', kind='high_ci':
+                    high confidence interval boundary for initialized skill
+                * result='p', kind='uninit':
+                    p value of the hypothesis that the
                     difference of skill between the initialized and
                     uninitialized simulations is smaller or equal to zero
                     based on bootstrapping with replacement.
-                * p_pers_over_init: p value of the hypothesis that the
+                * result='p', kind='pers':
+                    p value of the hypothesis that the
                     difference of skill between the initialized and persistence
                     simulations is smaller or equal to zero based on
                     bootstrapping with replacement.
@@ -990,37 +1022,45 @@ class HindcastEnsemble(PredictionEnsemble):
         alignment=None,
         **metric_kwargs,
     ):
-        """Verifies the initialized ensemble against observations/verification data.
+        """Verifies the initialized ensemble against observations.
 
-        This will automatically verify against all shared variables
-        between the initialized ensemble and observations/verification data.
+        .. note::
+            This will automatically verify against all shared variables
+            between the initialized ensemble and observations/verification data.
 
         Args:
             name (str): Short name of observations/verification data to compare to.
                 If ``None``, compare to all observations/verification data.
-            reference (str): Which reference skill is also computed. Choose
-                from ['historical', 'persistence']. Defaults to None.
-            metric (str): Metric to apply for verification.
-            comparison (str, default 'e2o'): How to compare to the
-                observations/verification data. ('e2o' for ensemble mean to
-                observations/verification data. 'm2o' for each individual member to
-                observations/verification data).
-            dim (str, list of str): dimension(s) to apply metric over.
+            reference (str): Type of reference forecasts to also verify against the
+                observations. Choose one or more of ['historical', 'persistence'].
+                Defaults to None.
+            metric (str): Metric to apply for verification. see
+                `metrics </metrics.html>`_.
+            comparison (str): How to compare to the
+                observations/verification data. See `comparisons </comparisons.html>`_.
+            dim (str, list of str): Dimension(s) to apply metric over. ``dim`` is passed
+                on to xskillscore.{metric} and includes xskillscore's ``member_dim``.
+                ``dim`` should contain ``member`` when ``comparison`` is probabilistic
+                but should not contain ``member`` when ``comparison=e2o``. Defaults to
+                ``None`` meaning that all dimensions other than ``lead`` are reduced.
             alignment (str): which inits or verification times should be aligned?
-                - maximize/None: maximize the degrees of freedom by slicing ``hind`` and
+
+                - 'maximize': maximize the degrees of freedom by slicing ``hind`` and
                 ``verif`` to a common time frame at each lead.
-                - same_inits: slice to a common init frame prior to computing
+                - 'same_inits': slice to a common init frame prior to computing
                 metric. This philosophy follows the thought that each lead should be
                 based on the same set of initializations.
-                - same_verif: slice to a common/consistent verification time frame prior
-                to computing metric. This philosophy follows the thought that each lead
-                should be based on the same set of verification dates.
-            **metric_kwargs (optional): arguments passed to `metric`.
+                - 'same_verif': slice to a common/consistent verification time frame
+                prior to computing metric. This philosophy follows the thought that
+                each lead should be based on the same set of verification dates.
+
+            **metric_kwargs (optional): arguments passed to ``metric``.
 
         Returns:
             Dataset of comparison results (if comparing to one observational product),
             or dictionary of Datasets with keys corresponding to
             observations/verification data short name.
+
         """
         if isinstance(reference, str):
             reference = [reference]
@@ -1149,43 +1189,61 @@ class HindcastEnsemble(PredictionEnsemble):
         pers_sig=None,
         **metric_kwargs,
     ):
-        """Bootstrap ensemble HindcastEnsemble with replacement according to
-        Goddard et al. 2013.
+        """Bootstrap with replacement according to Goddard et al. 2013.
 
         Args:
-            metric (str): Metric to apply for verification.
+            metric (str): Metric to apply for verification, see
+                `metrics </metrics.html>`_.
             comparison (str): How to compare to the
-                observations/verification data. ('e2o' for ensemble mean to
-                observations/verification data. 'm2o' for each individual member to
-                observations/verification data).
+                observations/verification data, see `comparisons </comparisons.html>`_.
+            dim (str, list of str): dimension(s) to apply metric over. ``dim`` is passed
+                on to xskillscore.{metric} and includes xskillscore's ``member_dim``.
+                ``dim`` should contain ``member`` when ``comparison`` is probabilistic
+                but should not contain ``member`` when ``comparison='e2o'``. Defaults to
+                ``None`` meaning that all dimensions other than ``lead`` are reduced.
             alignment (str): which inits or verification times should be aligned?
-                - maximize/None: maximize the degrees of freedom by slicing ``hind`` and
-                ``verif`` to a common time frame at each lead.
-                - same_inits: slice to a common init frame prior to computing
+
+                - 'maximize': maximize the degrees of freedom by slicing ``init`` and
+                ``verif`` to a common time frame at each lead
+                - 'same_inits': slice to a common init frame prior to computing
                 metric. This philosophy follows the thought that each lead should be
-                based on the same set of initializations.
-                - same_verif: slice to a common/consistent verification time frame prior
-                to computing metric. This philosophy follows the thought that each lead
-                should be based on the same set of verification dates.
-            sig (int, default 95):
-                Significance level for uninitialized and initialized
-                comparison.
-            dim (str): dimension to apply metric over. default: 'init'.
+                based on the same set of initializations
+                - 'same_verif': slice to a common/consistent verification time frame
+                prior to computing metric. This philosophy follows the thought that
+                each lead should be based on the same set of verification dates
+
+            sig (int, default 95): Significance level in percent for deciding whether
+                uninitialized and persistence beat initialized skill.
             resample_dim (str or list): dimension to resample from. default: 'member'.
 
                 - 'member': select a different set of members from hind
                 - 'init': select a different set of initializations from hind
 
             iterations (int, default 5): Number of resampling iterations for
-                bootstrapping with replacement. Recommended: 500.
+                bootstrapping with replacement. Recommended >=500.
             pers_sig (int, default None):
                 If not None, the separate significance level for persistence.
-            **metric_kwargs (optional): arguments passed to `metric`.
+            **metric_kwargs (optional): arguments passed to ``metric``.
 
         Returns:
-            Dataset of comparison results (if comparing to one observational product),
-            or dictionary of Datasets with keys corresponding to
-            observations/verification data short name.
+            xr.Datasets: with dimensions ``result`` (holding ``skill``, ``p``,
+            ``low_ci`` and ``high_ci``) and ``kind`` (holding ``init``, ``pers`` and
+            ``uninit``):
+                * result='skill', kind='init':
+                    mean initialized skill
+                * result='skill', kind='high_ci':
+                    high confidence interval boundary for initialized skill
+                * result='p', kind='uninit':
+                    p value of the hypothesis that the
+                    difference of skill between the initialized and
+                    uninitialized simulations is smaller or equal to zero
+                    based on bootstrapping with replacement.
+                * result='p', kind='pers':
+                    p value of the hypothesis that the
+                    difference of skill between the initialized and persistence
+                    simulations is smaller or equal to zero based on
+                    bootstrapping with replacement.
+
         """
         # TODO: replace with more computationally efficient classes implementation
         return bootstrap_hindcast(
@@ -1207,6 +1265,7 @@ class HindcastEnsemble(PredictionEnsemble):
 
         Args:
             alignment (str): which inits or verification times should be aligned?
+
                 - maximize/None: maximize the degrees of freedom by slicing ``hind`` and
                 ``verif`` to a common time frame at each lead.
                 - same_inits: slice to a common init frame prior to computing
@@ -1215,6 +1274,7 @@ class HindcastEnsemble(PredictionEnsemble):
                 - same_verif: slice to a common/consistent verification time frame prior
                 to computing metric. This philosophy follows the thought that each lead
                 should be based on the same set of verification dates.
+
             how (str or list of str): what kind of bias reduction to perform. Select
                 from ['mean']. Defaults to 'mean'.
             cross_validate (bool): Use properly defined mean bias reduction function.